Split db.ts into domain modules with barrel re-export
Replaces the monolithic db.ts (1,925 lines) with 15 focused domain modules
under apps/backend/src/db/:

- types.ts: All shared interfaces and type aliases
- connection.ts: DB connection management, schema init/migrate
- raw-records.ts: Raw record storage
- time-series.ts: Time series CRUD, bucketed aggregation, statistics
- activities.ts: Activity CRUD and merge logic
- locations.ts: Location, Place, Named/Detected location CRUD
- tags.ts: Tag CRUD and programmatic tag detection
- productivity.ts: RescueTime data
- lab-results.ts: Lab results storage
- oauth.ts: OAuth token storage
- sync-state.ts: Sync state tracking
- health-connect.ts: Health Connect data processing
- settings.ts: User settings CRUD
- lastfm-rules.ts: Last.fm auto-tagging rules
- mcp-sessions.ts: MCP session persistence

A barrel index.ts re-exports everything, so all existing imports
from './db' and '../db' continue working unchanged.

Phase 3 of the backend refactoring plan.
