+++ apps/backend/src/db/activities.ts
@@ -1,0 +1,198 @@
+/**
+ * Activity CRUD operations and merge logic.
+ */
+import type { ActivityType } from '../schema'
+import { query } from './connection'
+import { buildDynamicUpdate, type UpdateEntry } from './dynamic-update'
+import { mapActivityRow } from './row-mappers'
+import type { Activity, ActivityUpdate } from './types'
+
+/**
+ * Merge overlapping activities of the same type.
+ *
+ * When the same activity is logged in multiple apps (e.g., Polar for HR data
+ * and Gravl for workout details), this function merges them into a single
+ * activity using the earliest start time and latest end time.
+ *
+ * Merge rules:
+ * - Activities are grouped by activityType
+ * - Activities overlap if: a1.endTime >= a2.startTime (or a1 has no endTime and a2 starts during a1's day)
+ * - Merged activity uses: earliest startTime, latest endTime
+ * - First activity's source and id are kept
+ * - First non-empty title is used
+ * - Notes are concatenated with newline
+ * - Data objects are merged (later values override earlier for same keys)
+ */
+// eslint-disable-next-line complexity -- TODO: refactor
+export const mergeOverlappingActivities = (activities: Activity[]): Activity[] => {
+  if (activities.length === 0) return []
+
+  // Group by activity type
+  const byType = new Map<string, Activity[]>()
+  for (const a of activities) {
+    const group = byType.get(a.activityType) ?? []
+    group.push(a)
+    byType.set(a.activityType, group)
+  }
+
+  const result: Activity[] = []
+
+  for (const [, typeActivities] of byType) {
+    // Sort by start time
+    const sorted = [...typeActivities].sort((a, b) => a.startTime.getTime() - b.startTime.getTime())
+
+    let current = { ...sorted[0] }
+
+    for (let i = 1; i < sorted.length; i++) {
+      const next = sorted[i]
+      const currentEnd = current.endTime?.getTime() ?? current.startTime.getTime()
+      const nextStart = next.startTime.getTime()
+
+      // Check if activities overlap or touch
+      if (currentEnd >= nextStart) {
+        // Merge: extend end time if needed
+        const nextEnd = next.endTime?.getTime()
+        if (nextEnd !== undefined && (current.endTime === undefined || nextEnd > current.endTime.getTime())) {
+          current.endTime = next.endTime
+        }
+
+        // Use first non-empty title
+        if (!current.title && next.title) {
+          current.title = next.title
+        }
+
+        // Concatenate notes
+        if (next.notes) {
+          current.notes = current.notes ? `${current.notes}\n${next.notes}` : next.notes
+        }
+
+        // Merge data objects
+        if (next.data) {
+          current.data = { ...current.data, ...next.data }
+        }
+      } else {
+        // No overlap, save current and start new
+        result.push(current)
+        current = { ...next }
+      }
+    }
+
+    // Don't forget the last one
+    result.push(current)
+  }
+
+  // Sort final result by start time
+  return result.sort((a, b) => a.startTime.getTime() - b.startTime.getTime())
+}
+
+export const insertActivity = async (user: string, activity: Activity) => {
+  await query(
+    user,
+    `INSERT INTO activities (id, source, activity_type, start_time, end_time, title, notes, data)
+     VALUES (COALESCE($1, gen_random_uuid()), $2, $3, $4, $5, $6, $7, $8)
+     ON CONFLICT (source, activity_type, start_time) DO UPDATE SET
+       end_time = EXCLUDED.end_time,
+       title = EXCLUDED.title,
+       notes = EXCLUDED.notes,
+       data = EXCLUDED.data`,
+    [
+      activity.id,
+      activity.source,
+      activity.activityType,
+      activity.startTime,
+      activity.endTime,
+      activity.title,
+      activity.notes,
+      activity.data,
+    ],
+  )
+}
+
+export const getActivityById = async (user: string, id: string): Promise<Activity | null> => {
+  const result = await query(
+    user,
+    `SELECT id, source, activity_type, start_time, end_time, title, notes, data
+     FROM activities
+     WHERE id = $1`,
+    [id],
+  )
+
+  if (result.rows.length === 0) {
+    return null
+  }
+
+  return mapActivityRow(result.rows[0])
+}
+
+export const deleteActivity = async (user: string, id: string): Promise<boolean> => {
+  const result = await query(user, `DELETE FROM activities WHERE id = $1`, [id])
+
+  return (result.rowCount ?? 0) > 0
+}
+
+export const updateActivity = async (
+  user: string,
+  id: string,
+  updates: ActivityUpdate,
+): Promise<Activity | null> => {
+  const fields: UpdateEntry[] = []
+  if (updates.startTime !== undefined) fields.push({ column: 'start_time', value: updates.startTime })
+  if (updates.endTime !== undefined) fields.push({ column: 'end_time', value: updates.endTime })
+  if (updates.title !== undefined) fields.push({ column: 'title', value: updates.title })
+  if (updates.notes !== undefined) fields.push({ column: 'notes', value: updates.notes })
+
+  if (fields.length === 0) return getActivityById(user, id)
+
+  const update = buildDynamicUpdate('activities', id, fields, {
+    returning: 'id, source, activity_type, start_time, end_time, title, notes, data',
+  })
+  if (!update) return getActivityById(user, id)
+
+  const result = await query(user, update.sql, update.params)
+  if (result.rows.length === 0) return null
+  return mapActivityRow(result.rows[0])
+}
+
+export const getActivities = async (
+  user: string,
+  activityType: ActivityType | ActivityType[],
+  start: Date,
+  end: Date,
+): Promise<Activity[]> => {
+  const types = Array.isArray(activityType) ? activityType : [activityType]
+
+  const result = await query(
+    user,
+    `SELECT id, source, activity_type, start_time, end_time, title, notes, data
+     FROM activities
+     WHERE activity_type = ANY($1) AND start_time >= $2 AND start_time <= $3
+     ORDER BY start_time`,
+    [types, start, end],
+  )
+
+  const activities = result.rows.map(mapActivityRow)
+
+  return mergeOverlappingActivities(activities)
+}
+
+/**
+ * Get sleep sessions that overlap with a date range.
+ * Uses date overlap logic so overnight sleep (starting 11pm, ending 7am)
+ * appears on the wake-up day rather than the start day.
+ */
+export const getSleepSessions = async (user: string, start: Date, end: Date): Promise<Activity[]> => {
+  const result = await query(
+    user,
+    `SELECT id, source, activity_type, start_time, end_time, title, notes, data
+     FROM activities
+     WHERE activity_type = 'sleep'
+       AND start_time < $2
+       AND (end_time >= $1 OR end_time IS NULL)
+     ORDER BY start_time`,
+    [start, end],
+  )
+
+  const activities = result.rows.map(mapActivityRow)
+
+  return mergeOverlappingActivities(activities)
+}
