+++ apps/backend/src/db/connection.ts
@@ -1,0 +1,126 @@
+/**
+ * Database connection management and schema initialization.
+ */
+import { Client, type QueryResultRow } from 'pg'
+import format from 'pg-format'
+import { createTableStatements, tableCreationOrder } from '../schema'
+
+const dbByUser: Record<string, Client> = {}
+
+const userDbName = (user: string) => `aurboda_${user}`
+
+/**
+ * Inject a database client for a user. Used for testing with testcontainers.
+ * @internal
+ */
+export const _setClientForUser = (user: string, client: Client) => {
+  dbByUser[user] = client
+}
+
+export const query = async <T extends QueryResultRow = QueryResultRow>(
+  dbOrUser: Client | string,
+  queryStr: string,
+  params?: unknown[],
+) => {
+  const db = typeof dbOrUser === 'string' ? await getDbForUser(dbOrUser) : dbOrUser
+  const result = await db.query<T>(queryStr, params)
+  return result
+}
+
+export const loginToUserDb = async (user: string, password: string) => {
+  // Check if we already have a connection for this user
+  const existing = dbByUser[user]
+  if (existing) {
+    // Already connected - auth is handled by tokens, no need to re-verify password
+    // This avoids storing passwords in memory while maintaining security via token auth
+    return
+  }
+
+  const database = userDbName(user)
+  const client = new Client({ database, password, user })
+  await client.connect()
+  dbByUser[user] = client
+}
+
+export const makeNewUserDb = async (adminClient: Client, user: string, password: string) => {
+  const database = userDbName(user)
+  console.log(`New user ${user}`)
+  await query(adminClient, format('CREATE USER %I WITH ENCRYPTED PASSWORD %L', user, password))
+  await query(adminClient, format('GRANT %I TO %I', user, process.env.PGUSER))
+  await query(adminClient, format('CREATE DATABASE %I OWNER %I', database, user))
+
+  // Connect to the new database to create PostGIS extension (requires superuser privileges)
+  // We need a separate connection because CREATE EXTENSION operates on the current database
+  const newDbClient = new Client({ database })
+  await newDbClient.connect()
+  await query(newDbClient, 'CREATE EXTENSION IF NOT EXISTS postgis')
+  await newDbClient.end()
+
+  const client = new Client({ database, password, user })
+  await client.connect()
+  dbByUser[user] = client
+  await initializeSchema(user)
+}
+
+export const getDbForUser = async (user: string) => {
+  if (dbByUser[user]) return dbByUser[user]
+  const client = new Client({ database: userDbName(user) })
+  await client.connect()
+  await query(client, format('SET ROLE %L', user))
+  dbByUser[user] = client
+  return client
+}
+
+/**
+ * Initialize the database schema for a user.
+ * Creates all tables and indexes if they don't exist.
+ * Note: PostGIS extension is created in makeNewUserDb before this is called.
+ */
+export const initializeSchema = async (user: string) => {
+  const db = await getDbForUser(user)
+
+  for (const key of tableCreationOrder) {
+    await query(db, createTableStatements[key])
+  }
+}
+
+/**
+ * Run database migrations for a user.
+ * Checks which tables exist and creates missing ones.
+ */
+export const migrateSchema = async (user: string) => {
+  const db = await getDbForUser(user)
+  const database = `aurboda_${user}`
+
+  // Check which tables exist
+  const existingTables = await query(
+    db,
+    `SELECT table_name FROM information_schema.tables WHERE table_catalog = $1 AND table_schema = 'public'`,
+    [database],
+  )
+  const existingTableNames = new Set(existingTables.rows.map((r) => r.table_name))
+
+  // Create missing tables
+  for (const key of tableCreationOrder) {
+    const tableName = key.replace('_indexes', '')
+    if (!existingTableNames.has(tableName) || key.endsWith('_indexes')) {
+      // Always run index creation (IF NOT EXISTS handles duplicates)
+      // Create tables only if they don't exist
+      await query(db, createTableStatements[key])
+    }
+  }
+}
+
+/**
+ * Check if schema is initialized (has required tables).
+ */
+export const schemaInitialized = async (user: string) => {
+  const database = userDbName(user)
+  const db = await getDbForUser(user)
+  const result = await query(
+    db,
+    `SELECT 1 FROM information_schema.tables WHERE table_catalog = $1 AND table_name = $2`,
+    [database, 'raw_records'],
+  )
+  return result.rowCount !== 0
+}
